+++ server/src/routes/api/v1/bot/integration/index.ts
@@ -2,8 +2,6 @@ import { FastifyPluginAsync } from "fastify";
 import {
   createIntergationHandler,
   pauseOrResumeIntergationHandler,
-  whatsappIntergationHandler,
-  whatsappIntergationHandlerPost,
 } from "./handlers/post.handler";
 import {
   createIntergationSchema,
@@ -14,6 +12,11 @@ import {
 } from "./schema";
 
 import {
+  whatsappIntergationHandler,
+  whatsappIntergationHandlerPost,
+} from "./handlers/whatsapp.handler";
+
+import {
   generateAPIKeyHandler,
   getAPIIntegrationHandler,
   regenerateAPIKeyHandler,
@@ -22,42 +25,66 @@ import { getChannelsByProvider } from "./handlers/get.handler";
 
 const root: FastifyPluginAsync = async (fastify, _): Promise<void> => {
   // create integration for channel
-  fastify.post("/:id", {
-    schema: createIntergationSchema,
-    onRequest: [fastify.authenticate],
-  }, createIntergationHandler);
+  fastify.post(
+    "/:id",
+    {
+      schema: createIntergationSchema,
+      onRequest: [fastify.authenticate],
+    },
+    createIntergationHandler
+  );
   // pause or resume integration
-  fastify.post("/:id/toggle", {
-    schema: pauseOrResumeIntergationSchema,
-    onRequest: [fastify.authenticate],
-  }, pauseOrResumeIntergationHandler);
+  fastify.post(
+    "/:id/toggle",
+    {
+      schema: pauseOrResumeIntergationSchema,
+      onRequest: [fastify.authenticate],
+    },
+    pauseOrResumeIntergationHandler
+  );
 
   // return all bot channels
-  fastify.get("/:id", {
-    onRequest: [fastify.authenticate],
-  }, getChannelsByProvider);
+  fastify.get(
+    "/:id",
+    {
+      onRequest: [fastify.authenticate],
+    },
+    getChannelsByProvider
+  );
 
   // whatsapp integration
   fastify.get("/:id/whatsapp", {}, whatsappIntergationHandler);
   fastify.post("/:id/whatsapp", {}, whatsappIntergationHandlerPost);
 
   // api key integration
-  fastify.get("/:id/api", {
-    schema: getAPIIntegrationSchema,
-    onRequest: [fastify.authenticate],
-  }, getAPIIntegrationHandler);
+  fastify.get(
+    "/:id/api",
+    {
+      schema: getAPIIntegrationSchema,
+      onRequest: [fastify.authenticate],
+    },
+    getAPIIntegrationHandler
+  );
 
   // generate api key
-  fastify.post("/:id/api", {
-    schema: generateAPIKeySchema,
-    onRequest: [fastify.authenticate],
-  }, generateAPIKeyHandler);
+  fastify.post(
+    "/:id/api",
+    {
+      schema: generateAPIKeySchema,
+      onRequest: [fastify.authenticate],
+    },
+    generateAPIKeyHandler
+  );
 
   // regenerate api key
-  fastify.put("/:id/api", {
-    schema: regenerateAPIKeySchema,
-    onRequest: [fastify.authenticate],
-  }, regenerateAPIKeyHandler);
+  fastify.put(
+    "/:id/api",
+    {
+      schema: regenerateAPIKeySchema,
+      onRequest: [fastify.authenticate],
+    },
+    regenerateAPIKeyHandler
+  );
 };
 
 export default root;
